v5.0 (72)

+++ CSE/AppIntents/DeleteQuickSE.swift
@@ -20,6 +20,10 @@ struct DeleteQuickSE: AppIntent, CustomIntentMigratedAppIntent {
     }
 
     func perform() async throws -> some IntentResult {
+        if !IntentSupport.isAllowedEditingSearchEngines() {
+            throw IntentSupport.CSEIntentError.notAllowedEditingSearchEngines
+        }
+        
         CSEDataManager.deleteQuickCSE(cseID)
         return .result()
     }
